// src/app/models/prospect-client.ts
export class ProspectClient {
    id: number;
    code: string;
    name: string;
    address: string;
    zipCode: string;
    phone: string;
    email: string;
    salesEngagementCode: string;
    salesEngagementName: string;
    salesManagementCode: string;
    salesManagementName: string;
    engagementBranchCode: string;
    engagementBranchName: string;
    managementBranchCode: string;
    managementBranchName: string;
    lastProspectDate: Date;
    status: string;
    isCompleted: boolean;
    isSuccess: boolean;
}